finishing auth module

+++ lib/ui/screens/authentication_screens/register_screen/sub_screens/register_has_register.dart
@@ -1,19 +1,56 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../../../config/config.dart';
 import '../../../../ui.dart';
 import '../widets/widgets.dart';
 
-
-class HasRegisteredScreen extends StatefulWidget {
+class HasRegisteredScreen extends ConsumerStatefulWidget {
   final Function() onDeleteRegistration;
   const HasRegisteredScreen({super.key, required this.onDeleteRegistration});
 
   @override
-  State<HasRegisteredScreen> createState() => _HasRegisteredScreenState();
+  ConsumerState<HasRegisteredScreen> createState() =>
+      _HasRegisteredScreenState();
 }
 
-class _HasRegisteredScreenState extends State<HasRegisteredScreen> {
+class _HasRegisteredScreenState extends ConsumerState<HasRegisteredScreen> {
+  int registrationAccepted = 0;
+
+  @override
+  void initState() {
+    // TODO: implement initState
+    super.initState();
+
+    // HelperPrefs.registrationRequest;
+    verifyIfRegistrationIsAcepted();
+  }
+
+  void verifyIfRegistrationIsAcepted() async {
+    final hasAcepted = await ref
+        .read(authUseCasesProvider)
+        .hasRegistrationAcepted(HelperPrefs.registrationRequest);
+    print('SOLICITUD ACEPTADA?');
+    print(hasAcepted);
+
+    switch (hasAcepted) {
+      case true:
+        setState(() {
+          registrationAccepted = 1;
+        });
+        break;
+      case false:
+        setState(() {
+          registrationAccepted = -1;
+        });
+        break;
+      default:
+        setState(() {
+          registrationAccepted = 0;
+        });
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
@@ -27,16 +64,26 @@ class _HasRegisteredScreenState extends State<HasRegisteredScreen> {
               Expanded(
                   flex: 3,
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.end,
+                    mainAxisAlignment: registrationAccepted == 1
+                        ? MainAxisAlignment.center
+                        : MainAxisAlignment.end,
                     children: [
                       Text(
-                        label1HasRegisteredUser,
+                        registrationAccepted == 1
+                            ? label1HasRegisteredUserAcepted
+                            : registrationAccepted == 0
+                                ? label1HasRegisteredUser
+                                : label1HasRegisteredUserNotAcepted,
                         style: theme.textTheme.labelMedium!
                             .copyWith(fontWeight: FontWeight.w800),
                         textAlign: TextAlign.center,
                       ),
                       Text(
-                        label2HasRegisteredUser,
+                        registrationAccepted == 1
+                            ? label2HasRegisteredUserAcepted
+                            : registrationAccepted == 0
+                                ? label2HasRegisteredUser
+                                : label2HasRegisteredUserNotAcepted,
                         style: theme.textTheme.labelMedium!,
                         textAlign: TextAlign.center,
                       ),
@@ -45,18 +92,30 @@ class _HasRegisteredScreenState extends State<HasRegisteredScreen> {
               const SizedBox(
                 height: 20,
               ),
-              Expanded(
-                  flex: 2,
-                  child: Align(
-                    alignment: Alignment.topCenter,
-                    child: CustomCircleButton(
-                      icon: Icons.close,
-                      onTap: () async {
-                        await HelperPrefs.clearPrefs();
-                        widget.onDeleteRegistration();
-                      },
-                    ),
-                  )),
+              registrationAccepted == 1
+                  ? SizedBox()
+                  : Expanded(
+                      flex: 2,
+                      child: Align(
+                        alignment: Alignment.topCenter,
+                        child: CustomCircleButton(
+                          icon: Icons.close,
+                          onTap: () async {
+                            final didDelete = await ref
+                                .read(authUseCasesProvider)
+                                .deleteRequest(HelperPrefs.registrationRequest);
+                            if (didDelete) {
+                              HelperPrefs.registrationRequest = '';
+                              widget.onDeleteRegistration();
+                              HelperNotificationUI.notificationSuccess(
+                                  'Solicitud eliminada');
+                            } else {
+                              HelperNotificationUI.notificationError(
+                                  'Su solicitud no se pudo eliminar, intentelo en otro momento por favor');
+                            }
+                          },
+                        ),
+                      )),
             ],
           ),
         ),
